[unified search] bundle optimization
