+++ src/plugins/unified_search/public/dataview_picker/data_view_selector.tsx
@@ -95,3 +95,7 @@ export const DataViewSelector = ({
     </Fragment>
   );
 };
+
+// React.lazy support
+// eslint-disable-next-line import/no-default-export
+export default DataViewSelector;
