+++ src/plugins/unified_search/public/dataview_picker/dataview_list.tsx
@@ -90,3 +90,7 @@ export function DataViewsList({
     </EuiSelectable>
   );
 }
+
+// React.lazy support
+// eslint-disable-next-line import/no-default-export
+export default DataViewsList;
