+++ src/plugins/unified_search/public/search_bar/lib/populate_state_from_saved_query.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { QueryStart, SavedQuery } from '@kbn/data-plugin/public';
+import type { QueryStart, SavedQuery } from '@kbn/data-plugin/public';
 
 export const populateStateFromSavedQuery = (queryService: QueryStart, savedQuery: SavedQuery) => {
   const {
