+++ src/plugins/unified_search/public/search_bar/lib/use_saved_query.ts
@@ -9,9 +9,9 @@
 import { useState, useEffect } from 'react';
 import { i18n } from '@kbn/i18n';
 
-import { CoreStart } from '@kbn/core/public';
-import { SavedQuery } from '@kbn/data-plugin/public';
-import { DataPublicPluginStart } from '@kbn/data-plugin/public';
+import type { CoreStart } from '@kbn/core/public';
+import type { SavedQuery } from '@kbn/data-plugin/public';
+import type { DataPublicPluginStart } from '@kbn/data-plugin/public';
 import { populateStateFromSavedQuery } from './populate_state_from_saved_query';
 import { clearStateFromSavedQuery } from './clear_saved_query';
