+++ src/plugins/unified_search/public/search_bar/lib/use_timefilter.ts
@@ -8,7 +8,7 @@
 
 import { useState, useEffect } from 'react';
 import { Subscription } from 'rxjs';
-import { DataPublicPluginStart, RefreshInterval } from '@kbn/data-plugin/public';
+import type { DataPublicPluginStart, RefreshInterval } from '@kbn/data-plugin/public';
 import type { TimeRange } from '@kbn/es-query';
 
 interface UseTimefilterProps {
